Updated gauge to also validate labels
