+++ src/clients/victoriaMetrics/promMetricCounter.ts
@@ -1,5 +1,9 @@
 import promClient, { Counter } from 'prom-client';
 import { promAgent } from './promAgent';
+import {
+  generateLabelCombinations,
+  validateLabelNamesAndOptions,
+} from './metricFunctions';
 
 export interface CounterConfig {
   name: string;
@@ -27,11 +31,11 @@ export class PromMetricCounter {
 
     // Validate labelNames and labelOptions if both provided
     if (counterConfig.labelNames && counterConfig.labelOptions) {
-      this.validateLabelNamesAndOptions(
+      validateLabelNamesAndOptions(
         counterConfig.labelNames,
         counterConfig.labelOptions,
       );
-      this.labelCombinations = this.generateLabelCombinations(
+      this.labelCombinations = generateLabelCombinations(
         counterConfig.labelOptions,
       );
     }
@@ -43,52 +47,6 @@ export class PromMetricCounter {
         this.counter.inc(combination, 0);
       });
     }
-  }
-
-  private validateLabelNamesAndOptions(
-    labelNames: string[],
-    labelOptions: Record<string, string[]>,
-  ) {
-    const labelNamesFromOptions = Object.keys(labelOptions);
-    const invalidLabelNamesFromOptions = labelNamesFromOptions.filter(
-      (labelName) => !labelNames.includes(labelName),
-    );
-
-    if (invalidLabelNamesFromOptions.length > 0) {
-      throw new Error(
-        `Error: Undeclared labelNames: "${invalidLabelNamesFromOptions}".`,
-      );
-    }
-
-    const invalidLabelNames = labelNames.filter(
-      (labelName) => !labelNamesFromOptions.includes(labelName),
-    );
-    if (invalidLabelNames.length > 0) {
-      throw new Error(
-        `Error: labelOptions are not configured for labelNames="${invalidLabelNames}".`,
-      );
-    }
-  }
-
-  private generateLabelCombinations(
-    labelOptions?: Record<string, string[]>,
-  ): Record<string, string>[] {
-    let labelCombinations = [{}];
-
-    Object.keys(labelOptions).forEach((labelName) => {
-      // Get current label options
-      const options = labelOptions[labelName];
-
-      // Update combinations with new options
-      labelCombinations = labelCombinations.flatMap((combination) =>
-        options.map((option) => ({
-          ...combination,
-          [labelName]: option,
-        })),
-      );
-    });
-
-    return labelCombinations;
   }
 
   public inc(num: number, labels?: Record<string, string>) {
